pyproject.toml
[tool.poetry]
name = "fs-schema-validator"
version = "0.1.0"
description = ""
authors = ["Federico Ravasio <fr@nablaflow.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"

click = "^8.1.3"
orjson = "^3.8.0"
parsita = "^1.7.1"
pillow = "^9.2.0"
pydantic = "^1.10.2"
pygltflib = "^1.15.3"
pyyaml = "^6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.8.0"
isort = "^5.10.1"
mypy = "^0.971"
pytest = "^7.1.3"
types-pillow = "^9.2.0"
types-pyyaml = "^6.0"

[tool.pytest.ini_options]
addopts = "-vv"
pythonpath = "."

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3

[tool.mypy]
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_decorators = true
disallow_untyped_defs = true
follow_imports = 'silent'
no_implicit_reexport = true
show_error_codes = true
strict_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "parsita",
    "pygltflib",
]
ignore_missing_imports = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
